20.4 We can navigate between the individual pages, and everything looks and behaves like it did before. We can even select one of these to delete. And whenever we confirm the deletion, we'll see that the results updates. That went from 26 to 25 results. And of course, the articles that we see inside of the page is different than what we had before.  Now, you might be thinking, what's the purpose then? The behavior, the functionality is all of the same. And besides, I thought we were going to use a computed property so that we can cache that information. Well, yes. But by default, computed properties last for a single request. And it's not just hitting the refresh button on the browser. It's any request, including live wire requests. So whenever I click on a delete button and then confirm that delete, that is issuing a request. So articles is then going to be pulled from the database, and those articles will then be used to render in the browser.  So really, once the articles are rendered in the browser, that's it. The cache is gone. However, we can make some changes so that we do see that the cache is actually being used. So whenever we delete an article, the first thing I want to do is access the articles here. So by doing it here inside of the delete action, we are fetching the articles and it is going to be cached. So let's do something like this. If the count of the articles is less than 10, then we don't want to delete. So we'll just throw a new exception and the message doesn't matter. We'll just say nope. And if this condition is false, then we will actually delete the article.  But bear in mind that by the time we delete the article, we still have a cached set of articles. And that cached set is going to be what is rendered inside of the page because, well, it's cached. So if we pick one of these to delete, and let's make sure that we are on a page that has at least nine items, and we'll just delete this one.  Pay attention to the status information showing 11 to 20 of 25 results because whenever I click on OK, it looks like that nothing happened. However, whenever I refresh the page, then we can see that, yes, there was an article deleted, and now we see the new list of results. Now, once again, we see this behavior because whenever we attempt to delete an article, we are accessing the articles. So this is the first time we access our articles. It's going to fetch those articles from the database. It's going to cache the set of articles that we are going to display so that whenever we delete the article, that doesn't break the cache in any way. That just changes what we have in the database. So then whenever we display it in the view, we still are displaying that cached set of articles.

+++ app/Livewire/ArticleList.php
@@ -23,6 +23,10 @@ class ArticleList extends AdminComponent {
 	}
 
 	public function delete(Article $article) {
+		if ($this->articles->count() < 10) {
+			throw new \Exception("Nope");
+		}
+		
 		$article->delete();
 	}
